python-generators-0x00/requirements.txt
mysql-connector-python==9.5.0
numpy==2.3.4
pandas==2.3.3

python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
tzdata==2025.2
mysql-connector-python==9.5.0
numpy==2.3.4
pandas==2.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
tzdata==2025.2
